feat: 多开ui优化
